# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ContactOut": {
        "properties": {
          "agency_score": {
            "title": "Agency Score",
            "type": "number"
          },
          "classification": {
            "title": "Classification",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "identifier": {
            "title": "Identifier",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "identifier",
          "display_name",
          "classification",
          "agency_score"
        ],
        "title": "ContactOut",
        "type": "object"
      },
...
      "DuplicateOut": {
        "properties": {
          "property_id": {
            "format": "uuid",
            "title": "Property Id",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "property_id",
          "score"
        ],
        "title": "DuplicateOut",
        "type": "object"
      },
...
      "ListingOut": {
        "properties": {
          "address_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Text"
          },
          "agent_marker": {
            "title": "Agent Marker",
            "type": "boolean"
          },
          "area_sqm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Sqm"
          },
          "contacts": {
            "items": {
              "$ref": "#/components/schemas/ContactOut"
            },
            "title": "Contacts",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "external_id": {
            "title": "External Id",
            "type": "string"
          },
          "first_seen_at": {
            "format": "date-time",
            "title": "First Seen At",
            "type": "string"
          },
          "floor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "owner_marker": {
            "title": "Owner Marker",
            "type": "boolean"
          },
          "parse_confidence": {
            "title": "Parse Confidence",
            "type": "number"
          },
          "photos": {
            "items": {
              "$ref": "#/components/schemas/PhotoOut"
            },
            "title": "Photos",
            "type": "array"
          },
          "posted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posted At"
          },
          "price": {
            "$ref": "#/components/schemas/PriceOut"
          },
          "rooms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rooms"
          },
          "source": {
            "$ref": "#/components/schemas/SourceRef"
          },
          "source_removed": {
            "title": "Source Removed",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "total_floors": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Floors"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "id",
          "source",
          "external_id",
          "url",
          "title",
          "description",
          "price",
          "rooms",
          "area_sqm",
          "floor",
          "total_floors",
          "district",
          "address_text",
          "posted_at",
          "first_seen_at",
          "last_seen_at",
          "source_removed",
          "owner_marker",
          "agent_marker",
          "parse_confidence",
          "photos",
          "contacts"
        ],
        "title": "ListingOut",
        "type": "object"
      },
...
      "OwnerOut": {
        "properties": {
          "agency_score": {
            "title": "Agency Score",
            "type": "number"
          },
          "classification": {
            "title": "Classification",
            "type": "string"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "contact_id": {
            "format": "uuid",
            "title": "Contact Id",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "identifier": {
            "title": "Identifier",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "contact_id",
          "kind",
          "identifier",
          "display_name",
          "classification",
          "agency_score",
          "confidence"
        ],
        "title": "OwnerOut",
        "type": "object"
      },
...
      "Problem": {
        "description": "The body of every error response, sent as `application/problem+json`.",
        "properties": {
          "code": {
            "description": "The machine-readable error code the web translates. One of:\n`auth.missing_token`, `auth.token_expired`, `auth.token_invalid`, `auth.user_inactive`,\n`auth.invalid_credentials`, `auth.forbidden`, `not_found`, `method_not_allowed`,\n`validation_error`, `internal_error`, `listing.unsupported_url`, `listing.invalid_url`,\n`listing.gone`, `source.misconfigured`, `source.unavailable`, `source.login_required`,\n`source.peer_unresolved`, `source.exists`; any other HTTP status raised by the framework\nbecomes `http.<status>`.",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
      "PropertyDetail": {
        "properties": {
          "area_sqm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Sqm"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "duplicates": {
            "items": {
              "$ref": "#/components/schemas/DuplicateOut"
            },
            "title": "Duplicates",
            "type": "array"
          },
          "first_seen_at": {
            "format": "date-time",
            "title": "First Seen At",
            "type": "string"
          },
          "floor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "last_status_event": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StatusEventOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "listing_count": {
            "title": "Listing Count",
            "type": "integer"
          },
          "listings": {
            "items": {
              "$ref": "#/components/schemas/ListingOut"
            },
            "title": "Listings",
            "type": "array"
          },
          "needs_recheck": {
            "title": "Needs Recheck",
            "type": "boolean"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "price_usd_min_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Usd Min Minor"
          },
          "probable_owner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OwnerOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "rooms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rooms"
          },
          "source_kinds": {
            "items": {
              "type": "string"
            },
            "title": "Source Kinds",
            "type": "array"
          },
          "source_removed": {
            "title": "Source Removed",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "status_events": {
            "items": {
              "$ref": "#/components/schemas/StatusEventOut"
            },
            "title": "Status Events",
            "type": "array"
          },
          "total_floors": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Floors"
          }
        },
        "required": [
          "id",
          "status",
          "district",
          "rooms",
          "floor",
          "total_floors",
          "area_sqm",
          "price_usd_min_minor",
          "source_removed",
          "needs_recheck",
          "first_seen_at",
          "last_seen_at",
          "listing_count",
          "source_kinds",
          "probable_owner",
          "photo_url",
          "last_status_event",
          "listings",
          "status_events",
          "duplicates"
        ],
        "title": "PropertyDetail",
        "type": "object"
      },
      "PropertyPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PropertyRow"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "PropertyPage",
        "type": "object"
      },
      "PropertyRow": {
        "properties": {
          "area_sqm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Sqm"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "first_seen_at": {
            "format": "date-time",
            "title": "First Seen At",
            "type": "string"
          },
          "floor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "last_status_event": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StatusEventOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "listing_count": {
            "title": "Listing Count",
            "type": "integer"
          },
          "needs_recheck": {
            "title": "Needs Recheck",
            "type": "boolean"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "price_usd_min_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Usd Min Minor"
          },
          "probable_owner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OwnerOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "rooms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rooms"
          },
          "source_kinds": {
            "items": {
              "type": "string"
            },
            "title": "Source Kinds",
            "type": "array"
          },
          "source_removed": {
            "title": "Source Removed",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "total_floors": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Floors"
          }
        },
        "required": [
          "id",
          "status",
          "district",
          "rooms",
          "floor",
          "total_floors",
          "area_sqm",
          "price_usd_min_minor",
          "source_removed",
          "needs_recheck",
          "first_seen_at",
          "last_seen_at",
          "listing_count",
          "source_kinds",
          "probable_owner",
          "photo_url",
          "last_status_event"
        ],
        "title": "PropertyRow",
        "type": "object"
      },
...
      "SourceRef": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "name"
        ],
        "title": "SourceRef",
        "type": "object"
      },
...
      "StatusEventOut": {
        "properties": {
          "actor_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Id"
          },
          "actor_type": {
            "title": "Actor Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "from_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Status"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "to_status": {
            "title": "To Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "from_status",
          "to_status",
          "actor_type",
          "actor_id",
          "note",
          "created_at"
        ],
        "title": "StatusEventOut",
        "type": "object"
      },
      "StatusIn": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "status": {
            "enum": [
              "active",
              "inactive"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "StatusIn",
        "type": "object"
      },
...
      "ValidationIssue": {
        "properties": {
          "loc": {
            "items": {
              "type": "string"
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationIssue",
        "type": "object"
      },
      "ValidationProblem": {
        "properties": {
          "code": {
            "description": "The machine-readable error code the web translates. One of:\n`auth.missing_token`, `auth.token_expired`, `auth.token_invalid`, `auth.user_inactive`,\n`auth.invalid_credentials`, `auth.forbidden`, `not_found`, `method_not_allowed`,\n`validation_error`, `internal_error`, `listing.unsupported_url`, `listing.invalid_url`,\n`listing.gone`, `source.misconfigured`, `source.unavailable`, `source.login_required`,\n`source.peer_unresolved`, `source.exists`; any other HTTP status raised by the framework\nbecomes `http.<status>`.",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Errors",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "detail",
          "code",
          "errors"
        ],
        "title": "ValidationProblem",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "realtor-app API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/properties": {
      "get": {
        "operationId": "list_properties_endpoint_api_v1_properties_get",
        "parameters": [
          {
            "in": "query",
            "name": "district",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District"
            }
          },
          {
            "in": "query",
            "name": "rooms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rooms"
            }
          },
          {
            "description": "whole USD",
            "in": "query",
            "name": "price_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "whole USD",
              "title": "Price Min"
            }
          },
          {
            "description": "whole USD",
            "in": "query",
            "name": "price_max",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "whole USD",
              "title": "Price Max"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "olx",
                    "telegram",
                    "manual"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "owner_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Owner Only",
              "type": "boolean"
            }
          },
          {
            "description": "include properties removed at the source",
            "in": "query",
            "name": "removed",
            "required": false,
            "schema": {
              "default": false,
              "description": "include properties removed at the source",
              "title": "Removed",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "last_seen",
              "enum": [
                "last_seen",
                "first_seen",
                "price_asc",
                "price_desc"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "missing, invalid or expired token; or inactive user"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "request validation failed"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Properties Endpoint",
        "tags": [
          "properties"
        ]
      }
    },
    "/api/v1/properties/{property_id}": {
      "get": {
        "operationId": "property_detail_api_v1_properties__property_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Property Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "missing, invalid or expired token; or inactive user"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "unknown property"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "request validation failed"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Property Detail",
        "tags": [
          "properties"
        ]
      }
    },
    "/api/v1/properties/{property_id}/status": {
      "post": {
        "operationId": "set_property_status_api_v1_properties__property_id__status_post",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Property Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusEventOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "missing, invalid or expired token; or inactive user"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "unknown property"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "request validation failed"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Property Status",
        "tags": [
          "properties"
        ]
      }
    },
...
  }
}